index.vueのスリム化（使用していないコンポーネント等の記述の削除）。表示しているカードをカードディレクトリへ移動。また、カードコンポーネントへ機能の集約。i18n対応は未対応。
