+++ pages/index.vue
@@ -1,41 +1,15 @@
 <template>
   <div class="MainPage">
     <page-header
-      :icon="headerItem.icon"
-      :title="headerItem.title"
-      :date="headerItem.date"
+      :icon="'mdi-chart-timeline-variant'"
+      :title="this.$t('都内の最新感染動向')"
+      :date="Data.lastUpdate"
     />
     <whats-new class="mb-4" :items="newsItems" />
     <v-row class="DataBlock">
-      <v-col cols="12" md="6" class="DataCard">
-        <confirmed-cases-details-card />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <time-bar-chart
-          :title="$t('陽性患者数')"
-          :title-id="'number-of-confirmed-cases'"
-          :chart-id="'time-bar-chart-patients'"
-          :chart-data="patientsGraph"
-          :date="Data.patients.date"
-          :unit="$t('人')"
-          :url="
-            'https://opendata.pref.saitama.lg.jp/data/dataset/covid19-jokyo'
-          "
-        />
-      </v-col>
-      <v-col cols="12" md="6" class="DataCard">
-        <data-table
-          :title="$t('陽性患者の属性')"
-          :title-id="'attributes-of-confirmed-cases'"
-          :chart-data="patientsTable"
-          :chart-option="{}"
-          :date="Data.patients.date"
-          :info="sumInfoOfPatients"
-          :url="
-            'https://opendata.pref.saitama.lg.jp/data/dataset/covid19-jokyo'
-          "
-        />
-      </v-col>
+      <confirmed-cases-details-card />
+      <confirmed-cases-number-card />
+      <confirmed-cases-attributes-card />
     </v-row>
   </div>
 </template>
@@ -281,169 +255,30 @@
 
 <script>
 import PageHeader from '@/components/PageHeader.vue'
-import ConfirmedCasesDetailsCard from '@/components/ConfirmedCasesDetailsCard.vue'
-import TimeBarChart from '@/components/TimeBarChart.vue'
 import WhatsNew from '@/components/WhatsNew.vue'
+import ConfirmedCasesDetailsCard from '@/components/cards/ConfirmedCasesDetailsCard.vue'
+import ConfirmedCasesNumberCard from '@/components/cards/confirmed-cases-number-card.vue'
+import ConfirmedCasesAttributesCard from '@/components/cards/confirmed-cases-attributes-card.vue'
 import Data from '@/data/data.json'
-import MetroData from '@/data/metro.json'
-import DataTable from '@/components/DataTable.vue'
-import formatGraph from '@/utils/formatGraph'
-import formatTable from '@/utils/formatTable'
 import formatConfirmedCases from '@/utils/formatConfirmedCases'
 import News from '@/data/news.json'
 
 export default {
   components: {
     PageHeader,
-    ConfirmedCasesDetailsCard,
-    TimeBarChart,
     WhatsNew,
-    DataTable
+    ConfirmedCasesDetailsCard,
+    ConfirmedCasesNumberCard,
+    ConfirmedCasesAttributesCard
   },
   data() {
-    // 感染者数グラフ
-    const patientsGraph = formatGraph(Data.patients_summary.data)
-    // 陽性患者の属性
-    const patientsTable = formatTable(Data.patients.data)
-    // 陽性患者の属性 ヘッダー翻訳
-    for (const header of patientsTable.headers) {
-      header.text =
-        header.value === '退院' ? this.$t('退院※') : this.$t(header.value)
-    }
-    // 陽性患者の属性 中身の翻訳
-    for (const row of patientsTable.datasets) {
-      row['居住地'] = this.$t(row['居住地'])
-      row['性別'] = this.$t(row['性別'])
-
-      if (row['年代'] === '10歳未満') {
-        row['年代'] = this.$t('10歳未満')
-      } else if (row['年代'].slice(-1) === '代') {
-        const age = row['年代'].substring(0, 2)
-        row['年代'] = this.$t('{age}代', { age })
-      }
-    }
-    // 退院者グラフ
-    const dischargesGraph = formatGraph(Data.discharges_summary.data)
-
-    // 相談件数
-    const contactsGraph = formatGraph(Data.contacts.data)
-    // 帰国者・接触者電話相談センター相談件数
-    const querentsGraph = formatGraph(Data.querents.data)
-    // 都営地下鉄の利用者数の推移
-    const metroGraph = MetroData
-    // metroGraph ツールチップ title文字列
-    // this.$t を使うため metroGraphOption の外側へ
-    const metroGraphTooltipTitle = (tooltipItems, _) => {
-      const label = tooltipItems[0].label
-      return this.$t('期間: {duration}', {
-        // duration = label = '2月10日~14日' | '2月17日~21日' | '2月25日~28日'
-        duration: this.$t(label)
-      })
-    }
-    // metroGraph ツールチップ label文字列
-    // this.$t を使うため metroGraphOption の外側へ
-    const metroGraphTooltipLabel = (tooltipItem, data) => {
-      const currentData = data.datasets[tooltipItem.datasetIndex]
-      const percentage = `${currentData.data[tooltipItem.index]}%`
-
-      return this.$t('{duration}の利用者数との相対値: {percentage}', {
-        // duration = metroGraph.base_period = '1月20日~1月24日'
-        duration: this.$t(metroGraph.base_period),
-        percentage
-      })
-    }
-    // 検査実施日別状況
-    const inspectionsGraph = [
-      Data.inspections_summary.data['都内'],
-      Data.inspections_summary.data['その他']
-    ]
-    const inspectionsItems = [
-      this.$t('都内発生（疑い例・接触者調査）'),
-      this.$t('その他（チャーター便・クルーズ船）')
-    ]
-    const inspectionsLabels = Data.inspections_summary.labels
-    // 死亡者数
-    // #MEMO: 今後使う可能性あるので一時コメントアウト
-    // const fatalitiesTable = formatTable(
-    //   Data.patients.data.filter(patient => patient['備考'] === '死亡')
-    // )
     // 検査陽性者の状況
     const confirmedCases = formatConfirmedCases(Data.main_summary)
 
-    const sumInfoOfPatients = {
-      lText: patientsGraph[
-        patientsGraph.length - 1
-      ].cumulative.toLocaleString(),
-      sText: this.$t('{date}の累計', {
-        date: patientsGraph[patientsGraph.length - 1].label
-      }),
-      unit: this.$t('人')
-    }
-
     const data = {
       Data,
-      patientsTable,
-      patientsGraph,
-      dischargesGraph,
-      contactsGraph,
-      querentsGraph,
-      metroGraph,
-      inspectionsGraph,
-      inspectionsItems,
-      inspectionsLabels,
       confirmedCases,
-      sumInfoOfPatients,
-      headerItem: {
-        icon: 'mdi-chart-timeline-variant',
-        title: this.$t('都内の最新感染動向'),
-        date: Data.lastUpdate
-      },
-      newsItems: News.newsItems,
-      metroGraphOption: {
-        responsive: true,
-        legend: {
-          display: true
-        },
-        scales: {
-          xAxes: [
-            {
-              position: 'bottom',
-              stacked: false,
-              gridLines: {
-                display: true
-              },
-              ticks: {
-                fontSize: 10,
-                maxTicksLimit: 20,
-                fontColor: '#808080'
-              }
-            }
-          ],
-          yAxes: [
-            {
-              stacked: false,
-              gridLines: {
-                display: true
-              },
-              ticks: {
-                fontSize: 12,
-                maxTicksLimit: 10,
-                fontColor: '#808080',
-                callback(value) {
-                  return value.toFixed(2) + '%'
-                }
-              }
-            }
-          ]
-        },
-        tooltips: {
-          displayColors: false,
-          callbacks: {
-            title: metroGraphTooltipTitle,
-            label: metroGraphTooltipLabel
-          }
-        }
-      }
+      newsItems: News.newsItems
     }
     return data
   },
